feat: add in memory store
This can be used to store secrets in memory from containers so that its
not exposed on the host file system or in the container runtime itself.

This in memory store is accesiable via a simple rest api.
```bash
curl -X POST http://${CONTAINIFYCI_HOST}/mem/key -d value
curl http://${CONTAINIFYCI_HOST}/mem/key
```

The values in the in memory store can easily be accessed via defined
environment variables on the host. Like so:
```bash
export DOCKER_TOKEN="mem:accesstoken"
````

The in memory store is temporary.