fix: add progress logging format
The engine-ci tool has a new flag `--progress` that will enable progress
logging format. This format is useful for non ci/cd systems where its
beneficial to see the progress of the build in a more aggregated human
readable format.

The default format is still the old one (plain), but the new format can
be enabled by passing the `--progress progress` flag.