feat: add host runtime
Together to docker and podman also add host runtime to run task on host
machine instead of containers. This is useful for debugging or to speed
up the task execution. But it is not recommended to use it in production
environment.

It requires that all the dependencies of task to be executed are
installed on the host machine.

The runtime can be set with CONTAINER_RUNTIME=host environment variable.