fix(python): add dummy release command for python
This dummy command is used for the Github Release workflow. It just
prints a log message that's all. Still it lets finish the Github release
workflow which creates release and a tag. But for example don't upload
any artifacts as it would for a golang repository with a gorelease file.

Later the dummy release command can be replaced with a proper python
release command that handles artifact uploads similar to goreleaser
maybe.